requirements.txt
Flask==2.3.3
PyMuPDF==1.23.8
Pillow==10.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
pdf2image==1.16.3
easyocr==1.7.0
python-magic==0.4.27
Werkzeug==2.3.7

# 额外依赖：二维码/条形码解析
pyzbar==0.1.9

